Revert file rename to unscramble diff
