+++ src/pacmod_game_control_exec.cpp
@@ -5,7 +5,7 @@
  * See file LICENSE included with this software or go to https://opensource.org/licenses/MIT for full license details.
  */
 
-#include "pacmod_game_control/pacmod_game_control_node.h"
+#include "pacmod_game_control/pacmod_game_control.h"
 
 #include <memory>
